#111 A little bit more restructuring
