--- extension/ts/background/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"strict": false,
-		"moduleResolution": "node",
-		"target": "ES2015",
-		"types": [
-			"chrome"
-		],
-		"lib": [
-			"ES2021"
-		]
-	}
-}
