+++ extension/ts/background/workers/authorize.ts
